First batch of test_tensorexpr changes
[ghstack-poisoned]
